scheduled_messages_popover: Move selected_send_later_timestamp.
