+++ web/src/compose.js
@@ -28,6 +28,7 @@ import * as people from "./people";
 import * as rendered_markdown from "./rendered_markdown";
 import * as resize from "./resize";
 import * as rows from "./rows";
+import * as scheduled_messages from "./scheduled_messages";
 import * as scheduled_messages_popover from "./scheduled_messages_popover";
 import * as sent_messages from "./sent_messages";
 import * as server_events from "./server_events";
@@ -198,7 +199,7 @@ export function clear_compose_box() {
     compose_banner.clear_warnings();
     compose_banner.clear_uploads();
     compose_ui.hide_compose_spinner();
-    scheduled_messages_popover.reset_selected_schedule_timestamp();
+    scheduled_messages.reset_selected_schedule_timestamp();
 }
 
 export function send_message_success(local_id, message_id, locally_echoed) {
@@ -587,8 +588,7 @@ export function initialize() {
         )} .main-view-banner-action-button`,
         (event) => {
             event.preventDefault();
-            const send_at_timestamp =
-                scheduled_messages_popover.get_selected_send_later_timestamp();
+            const send_at_timestamp = scheduled_messages.get_selected_send_later_timestamp();
             scheduled_messages_popover.do_schedule_message(send_at_timestamp);
         },
     );
@@ -795,9 +795,8 @@ export function initialize() {
 function schedule_message_to_custom_date() {
     const compose_message_object = create_message_object();
 
-    const deliver_at = scheduled_messages_popover.get_formatted_selected_send_later_time();
-    const scheduled_delivery_timestamp =
-        scheduled_messages_popover.get_selected_send_later_timestamp();
+    const deliver_at = scheduled_messages.get_formatted_selected_send_later_time();
+    const scheduled_delivery_timestamp = scheduled_messages.get_selected_send_later_timestamp();
 
     const message_type = compose_message_object.type;
     let req_type;
